long not int

+++ ProzonPlugin/Config.cs
@@ -9,6 +9,6 @@ namespace ProzonPlugin
     public class Config
     {
         public string FolderPath = "Put a path here direct to the /prefabs folder";
-        public int ModId = 504209260;
+        public long ModId = 504209260;
     }
 }
